hot: fixing PCBA SN parsing
There was a small bug on detecting a Seagate drive and checking for PCBA SN that has been fixed. The original string had an ending space, whereas the comparison was missing this. Used strlen("SEAGATE") instead with strncmp to resolve this in case other code removes spaces or keeps them in the future.