hot: Disabling printing out MLU information
Commented out printing out what is looked up from MLU and MLUS. Right now this is not clear enough to output a message to properly warn at this level.
I left the messages and other code in place since it could still be useful.
Currently, these MLU fields have more meaning to whether the device queue or other commands may cause changes to sense data or resets, which is not the most useful at this time. Future changes will likely clear this up, but those must be approved by the T10 committee before they are implemented in this code.