feat: Cleaning up transition power
Cleaning up the transition power options to have a SAS/SATA option capable of EPC and legacy modes and a separate NVMe function for switching NVMe power states. This makes it more clear that they work differently. Also, NVMe can be handled in the SAS/SATA function with SCSI translation.