make: Turning on extra warnings in VS2017 projects
Turning on /Wall warnings to help catch more potential problems.
The following warnings are disabled:
4214;4201;4668;4820;4710;4255;5045;4711;4324;4221;4204

The reasons these are disabled vary slightly. For the most part, these are disabled because they will always show up or are not very useful, or only show up in Windows API headers.
Some warnings are disabled because they are for features we are using and are supported by other compilers that we use: nameless structs/unions.
Overall, disabling these will not be detremental to the objective of catching more issues as there are far more warnings enabled now than have been disabled.