feat: Adding new functions to assist with repopulating elements
Adding functions to send the Restore Elements and Rebuild commands. These repopulation commands are used to restore a drive to original state if the element can be restored.
The Get and Show physical elements functions also were updated to reflect a new bit stating whether or not the elements could be restored.