bug: Adding new definitions for max and infinite timeouts
Adding definitions for MAX and INFINITE timeout. Some OS's support infinite, and others don't. So there is a function that can be used to check for this capability as well.
For some commands, really long, or infinite timeouts max be necessary, so these are in use for those commands in order to assist with the ability to run them for different options within the tools.