feat: Adding code to handle pulling SMART vs GPL logs for certain log numbers
Handling pulling and listing logs via SMART and GPL directories. Adding some known log numbers to use when determining which command to use to pull them, while still defaulting the generic log pull to using GPL as that is where all current logs are going.

Also adding a warning output when a drive is responding in a way that it is not quite matching what is expected according to the specifications. This is more helpful debug in case we need to add additional workarounds to drive specific problems in the future.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>