feat: Adding an additional warning while format/fast format is reporting 100%
Adding a warning for format/fast format for when progress is reporting 100% but the sense data is still reporting that formatting is in progress. This was observed happening on some drives and waiting just a little longer allows enough time for it to finish up.
This would happen when performing a format/fast format without the poll for progress. When polling, it would stay in the loop until the sense data no longer shows format in progress so it is necessary for utilities to be able to show this warning to the users to prevent formatting errors.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>