bug: Fixing memory overrun issue when changing SAS phy speed settings
Fixing a bug where while trying to update the SAS phy's to a new speed, the code will perform an unnecessary reallocation and will additionally try to access more memory than was allocated. This causes a crash from reading or writing beyond the allocated memory. The issue was due to how the total memory that was allocated was being calculated and used when performing either a followup mode sense, as well as the mode select to make the changes.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>