bug: Changing how SCSI device write same support is checked
Changing how to check for write same support on SCSI devices. THis is complicated since there is no straight forward way to say in many cases.
This will likely require more adjustment to get this dialed in, but is a good improvement from the brokenness that was previously doing this check.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>