feat: Adding new FWDL flag for legacy drive updates. Also structure updates.
Adding a separate flag for ignoring the final segment of a download instead of keeping it built-in since this keeps causing new issues. This way it's selectable in case we are updating old drives.

Also, updated the FWDL structures to store version and size information in an attempt to improve future forward/backward compatibility.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>