quick: Making use of M_ToBool macro
Making use of the M_ToBool macro in many places in the code.
This macro was added to ensure no warnings within C++ for places in C that are common to check for some condition to assign to a bool, which produces a warning in C++, but not C. It's mostly for compatibility and cleanup.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>