clean: Used clang in VS2019 to check for additional warnings and fix them
Used the clang compiler in VS2019 to get additional warnings in the code and fixed a lot of the warnings that were issued. The vast majority were style related or warning about sign conversion which meant adding casts.
For the most part, there was nothing major here and there are still other warngings to resolve, but those are also style related warnings.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>