feat: Adding support for strange ATA passthrough case on FWDL
Adding setting the default transfer size to 512B on Windows if ATA passthrough is present and the FWDL API is NOT supported on that driver.
This is a workaround that goes back to Windows XP days where ATA passthrough can only do 512B transfers. This is not ideal, but should solve a recently reported issue. This check should be enhanced later when we have more information on specific drivers and specific versions affected by this issue, but for now this only sets this on Windows ATA passthrough for drivers lacking the firmware update IOCTL support.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>