bug: Checking in a few fixes necessary for fractional power consumption values.
Checking in fixes for checking and setting power consumptions values that can be reported as milliwatts or microwatts.
Until now only watts have been reported, so this bug was not noticed.
Currently these fixes work, but additional updates are needed to output 13.5w instead of rounding up to 14w so that customers know a more precise value that the drive supports for power consumption.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>