bug: erasing MBR before changing sector size to prevent weird issues
Adding in erasing the MBR before performing a fast format to ensure that the drive partitions can be reconfigured after changing the sector size.
On GPT formatted disks, a dummy MBR is written to ensure that old OSs do not think the drive is unformatted, so they look like a partition that extends to the end of the drive. In Windows, this shows up weird and will not allow for recreating the partitions since it implements this similar protection when it cannot find the GPT table (which is no longer in the correct sector after changing the sector size).
Since this operation has no guarantee of preserving data, an erase of the MBR is being added to ensure these weird issues do not happen in Windows going forward.

[Seagate/ToolBin#16]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>