clean: Fixing some warnings related to temperature output for hybrid SMART
Fixing warnings about truncation produced from the hybrid output for temperatures.
Since the values are int16_t, they have a max width of 6 including a minus sign. Because of this, it appears like it could be truncated, when this will never actually happen unless the drive firmware is reporting garbage.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>