feat: Calling function to unmount file systems before erasing data
Calling functions to unmount the mounted file systems before performing an erase that will erase the boot sectors. This is done to prevent some weird issues on some systems with certain types of partitions.
On Linux with ext4, somethign remained cached even after rereading the partition table that made it more difficult to repartition after erasing. This resolves the issue by unmounting the partitions before performing the erase.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>