bug: Fixing indicator for current block size in supported formats on USB
Fixing a bug in how the current block size is determined on USB device.
This is a bit complicated overall, but after doing a set sector size, the child drive blocksize must be checked for USB devices as the SATL will not necessarily be able to be refreshed to update it's internal cached information about the drive.

[Seagate/openSeaChest#79]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>