bug: Fixing detection of read-look ahead support in SCSI info
Fixing the detection for read-look ahead in SCSI reported information.
The part of the code that checks for support if it is not already enabled was not interpreting the changeable bit correctly. In changeable reporting, a 1 means "supported" but the logic being used was the same as with detection of the current state which is inverted due to how that bit reports the state (reports as 1 = disabled).

[Fixes Seagate/openSeaChest#85]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>