feat: Adding format corrupt message to features list on SAS
Adding a message when a drive is format corrupt, or at least reports format corrupt that not all features will be able to be identified. This has been observed in a few cases, so this will help provide some information that a user with a SCSI/SAS drive will need to format the drive before everything will be able to show up.

[Seagate/ToolBin#22]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>