feat: Adding log page size lookups for more standardized NVMe logs
Adding log page size lookups for more standardized NVMe logs. Many were missing in this old function or not returning a very useful size. This adds a lot, but there is still more to do for some logs that are variable length.
Some of the variable length logs are more complicated to calculate a length for than others which is why they are not part of this commit right now.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>