feat: Adding better NVMe log page sizes based on reading headers for length
Added reading the header to calculate the size of some NVMe log pages. This will improve support from the function that reports the log page sizes for a few additional logs.
The Supported Capacity Configuration List page is currently not supported by this function because determining the size is much more complicated. This can be revisited later.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>