feat: Adding automatic firmware update mode
Adding an automatic firmware update mode that will try to do deferred+activate whenever possible. This is more compatible with Windows 10 and later systems when it is supported by a drive. It will fallback to other modes if this is not supported by a drive. This makes this API easier to use for many people.

The structure size will not have changed but a new enum was made that preserves compatibility with the previous version, but adds new modes without interrupting behavior in opensea-transport.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>