bug: Rearranging ATA smart check code to catch some odd conditions
I recently encountered a product that confirms to the very old SFF-8035i spec in that the "return status" command is not supported...every other standard seems to require this command for the SMART feature set.
This code was capable of supporting this, but was not capturing the condition properly.
While working on that change, I also moved the backup SCSI smart check method to the end as a last resort. Between trying the return status command, reading the SCT status log, and comparing attributes and thresholds values, all of these other methods should capture any device that falls into this function and get a result.
The only reason this remains is if there is a device that is on an adapter not properly reporting result registers, does not support SCT status reporting, and does not support thresholds, then this final SCSI check is potentially useful if the translator support translating the SMART status. This is possible as there are some devices I've seen without the thresholds attached to USB bridges that do not report good status back....so it has potential for some value still, but is definitely a last resort.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>