clean: Cleaning up warnings from mingw/gcc builds
Cleaning up some warnings from some code changes to the firmware update code. GCC and MSCV disagree on case handling in one place and in another GCC warns about enum conversion, so casts were added. This was done since this enum conversion is at a place where backwards compatibility is being preserved when reading firmware feature support to report back to the caller if they are using an old version of the structure.
This is likely not an issue as the enums retained the same values for these modes to help ensure compatibility.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>