bug: Setting legacy ATA standby timer with idle in certain power modes
Using the drive's current power mode to determine whether to set the legacy ATA standby timer. Previously only standby was in use which was causing the drive to set the timer and spin down. Now it will use the idle command to configure the timer if it is NOT already in standby. If already in standby and changing the timer, it will use the standby command.

[Seagate/openSeaChest#107]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>