bug: Adding a special case for setting maxLBA when already at max for ATA
Adding a special case to setting the ATA maxLBA. Since both HPA and AMAC have some requirements around how these commands are processed between power cycles, it makes sense to NOT issue the command if there is not an actual change to be made right now since this uses up the one option to make a change in the current power cycle. While it is unlikely most people will hit this, it can still help when repeatedly testing/making changes.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>