bug: Implementing an automatic retry from deferred mode to segmented mode for FWDL
When the firmware download function is given a "automaic" option to figure out the best method to use, it will choose deferred download and activate when it can to be most compatible with current drives and operating systems.
Apparently some drives are reporting that they support deferred download mode even when they do not and reject the commands instead. So this retry checks for the specific conditions to do the retry in this case.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>