feat: Adding a new function to erase the boot sectors
Adding a new function to call to erase the boot sectors. If the low-level OS call can do it, that will be called first so that the writes from opensea-operations will succeed. I also added this call into the erase range function but it will only be run when the range writes to zero and the maxlba, otherwise it is not called.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>