doc: Adding output note on DST log that it is read in order
Adding note to the output of show DST log that the entries are read in order from newest to oldest according to the specifications for the drive. This was added because SATA reports a odd circular buffer that not all tools take the time to print in order from newest to oldest, so it may not look the same.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>