feat: Initial commit of code for reading partition table information
Initial code and make updates to read the partition table. Currently have code that works for MBR and GPT. Backup GPT is not yet supported or validated, but it will be added to wrap up GPT support.
Need to add APM support.

There are some API changes I also need to make for GPT to report the enum and/or the name of the partition type, but haven't figured out exactly how I want to make those changes yet.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>