feat: Adding code to read and dump the NVM power states in a human readable format
Adding a new set of functions to dump the supported NVMe power states in a more human friendly way. The specification is quite clear in what it reports, but this function simplifies it a bit to report the values all in milliwatts since the spec can report milli or centi watt values, making it a little easier to use this API since the units are uniform in these structures.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>