quick: Adding additional check so firmware activation will only be issued on successful FWDL
Adding an additional check to make sure the firmware activate command will only be issued if all the segment transfers occurred successfully. Not really a bug, just an excess command that would fail if the segments all failed.

I also added a memset before requesting some SCSI fwdl support to cleanup the verbose output since it was not necessarily zeroes between commands which could be misleading.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>