quick: Making change to mode select bytes for format unit
Changing format unit bytes to be more in line with other mode select changes.
Since a mode page is not being issued, just the header and block descriptor, the mode data length is being zeroed out now, as is the device specific parameter since those bits are either reserved or ignored during a mode select anyways.

I also adjusted the transfer sizes to be more correct with what is read and being sent back.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>