bug: Fixing setting enhanced ATA security erase bit for SAT security protocol commands
The SAT security erase unit buffer uses a different offset for the bit to perform enhanced security ersae so it is not a straight-forward memcpy like the other commands would allow.
This fixes setting the enhanced security erase bit according to the spec. It may be necessary to implement SATL unique workarounds in the future if any are found to translate this incorrectly.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>