feat: Adding a way to list the data sanitization capabilities of a given erase method
Different methods of erasing data from a drive qualify with differing levels of sanitization in the IEEE 2883 specification: clear & purge.

This change adds showing which erase qualifies with which level.
There is also a dump of information to inform the user what these differences mean.

There is a third case for clear, possible purge when the true level of the sanitization method cannot be confirmed (nvme format with secure erase) since this is important to let users know it will clear data and it might qualify as a purge.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>