bug: Fixing a couple ATA security bits and time output for completion
Fixing a couple bits that were in the wrong location for ATA security set password.
Fixing ATA security ZAC bit location...not currently used by utilities.

Fixing the time output when ATA security erase completes in under a second.
Completion in under a second can either be from a crypto erase or an error condition and this was not being handled.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

# Conflicts:
#	include/opensea_operation_version.h