bug: More fixes to ATA security bits
After another review and more testing fixing setting the correct bits in the correct location for both ATA passthrough and SAT ATA security translation.
Some bits were being set in the wrong place of the ATA word and others were fine.
They are all in the proper location and all have been test to set in the proper location according to the specifications for ATA and SAT (which can differ slightly in places)

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>

# Conflicts:
#	include/opensea_operation_version.h