feat: Disabling use of SAT ATA security protocol by default due to instability of HBA firmware
ATA security was implemented to use the SAT security protocol for ATA security when it was available to try and "play nice" with the HBAs that implemented it and prevent resets.
This does not actually happen in reality though. It is still just as likely to be interrupted.
Also, through testing over and over it is also unstable in the HBA firmware that says this protocol is supported and the commands will occasionally see the driver report errors and never issue the command.

So this change now only uses the SAT ATA security protocol for non-ATA drives (such as NVMe drives that implemented this protocol) or it will use SAT protocol when it is forced through booleans passed in.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>