feat: Specialized function to restore max LBA taking into account all ATA features
ATA has a lot of different features that can impact how max LBA is set and how to restore it.
HPA, HPA security extensions, DCO, and AMAC will all be handled by this option.
If any error occurs, then an error is reported to indicate that erase should not proceed until this has been corrected.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>