feat: Adding a function to help detect if HBA/bridge/adapter is in sync with drive capacity
Adding a feature to help detect if the adapter a device is attached to is in sync with the drive's change in capacity or not.
If a device is behind a SATL then it may not be up to date with the capacity once it has changed. The only known way to resolve this is to reboot the system.
This function will attempt reading VPD page 89h (SAT info) to see if that causes a refresh, but it is not guaranteed to refresh this information.
This can help a user determine that they need to do a reboot before they can use the drive at the new capacity.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>