feat: Adding in a reset to try synchronizing the setmaxlba change with a controller
Adding in issuing a reset to see if it helps synchronize a controller's knowledge of the attached drive's maxLBA value.
This seems to work on a broadcom controller in Linux and may work for others.
I don't love issuing the reset, but it's the only thing aside from a reboot that seems to work.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>