quick: Allowing "suspended" to continue running the offline routine
Allowing "suspended" to be treated as still in progress since this is what I observed just by running the offline routine on an old drive. This seems to make some sense as a way of indicating "this is still running, but not stopped" on the drive that I have to test.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>