feat: removing polling from the ATA SMART offline test
Removing the polling routine since testing various drives of different ages are all showing slightly different behavior when trying to poll and see if it is still in progress.

The polling is removed via a #define, so if we ever revisit this it could be enabled and debugged quickly as needed.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>