quick: Initial work to better support AUSE in sanitize
Initial work to better support the AUSE bit in sanitize commands.
Still need to convert this to a parameter, but that should be easier at this point.

The AUSE bit means "Allow unrestricted sanitize exit" which basically means, when this is set, the exit failure mode command can be sent to exit the failed status.
If this bit is NOT set, then it means the only way to clear a sanitize failure is to repeat the sanitize command until it passes.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>