feat: Add detection for ATA minor versions, ZAC versions, and zone domains/realms
Adding in support for detectiong the ATA and ZAC minor versions to report in the device info.
Also adding detection of the remove element and modify zones command.
Adding in some support to detect zone domains and zone realms support.
Add detecting ATA mutate and SCSI format with preset commands.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>