bug: Use a definition to use as the starting offset for ID Data
Using a new definition for where to begin the offset looping through the supported ID data log pages.
There was a fix a while back that corrected this, but there were a few other areas affected by not properly taking into account this starting offset and would fall short of reading the final page.
This definition is now used in all of these locations to correct this issue.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>