feat: Use a function to add to features and standards supported
Added new functions to handle appending to the list of supported features or supported standards by the device.
This cleaned up the code a bit and helps prevent a possible overflow if there are more features/standards supported than available space as this is checked prior to writing to these parts of the structures.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>