feat: Initial reorganization for ATA device information
Initial reorganization work to get ATA device information.
While working on handling adding features and standards, it was time to reorganize this into multiple functions to make it easier to reuse the code and reduce some duplication in some places.

There is still more work to do, but this is a big start to reorganizing this to make it easier to maintain long-term.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>