feat: Small optimization to reading log pages/VPD pages/mode pages
Making a small optimization to reading ATA log pages, SCSI VPD pages, log pages, and mode pages.
In all of these functions they were always issuing commands to determine the length needed to read all of the data.
This is ok in most cases, but in the cases where a buffer is provided with a length, it will use that length to fill in the data and skip the request to determine the size making a small performance improvement.
Also made a small improvement to handing old and RBC devices with the SCSI mode sense commands since the mode sense 6 command is only needed for these old devices. Retries are still available if mode sense 10 comes back with invalid operation code, but this small optimization may also improve performance.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>