feat: Adding in an optimization for pulling ATA ID log and dev stats logs
Adding in an optimization to skip over unsupported pages when reading the ATA identify device data log or the device statistics log.
Since these report the same first page with a list of supported pages, that is checked to determine when to skip issuing another read log ext command or not.
This is especially helpful for USB devices that read single sectors at a time on the device statistics log. When the vendor unique page is supported, there are a LOT of pages with nothing but zeroes that can be skipped, which leads to a performance improvement.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>