quick: Minor cleanup to ATA security while making sure ZAC bit is handled properly
Cleaned up ATA security code to use definitions from opensea-transport instead of a bunch of magic numbers.
The time printout for ATA security is also cleaned up with a function so that the duplicate code is reduced.
Also made sure the ZAC bit is handled correctly since that support is being added from openSeaChest_Erase.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>