feat: Adding initial code to take a requested capacity and output a resulting LBA value
Adding the initial code to calculate an LBA count based on a requested capacity.
This follows the methods in SFF-8447 to determine the LBA counts for 50GB - present capacities.
The function currently reuses a drive's current sector size and PI mode for the calculation since this is intended to be used for changing a drive to a lower capacity during a change to the maxLBA with HPA/AMAC/etc

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>