feat: Adding more sanitize info from NVMe and handling no-deallocate better
NVMe has more info regarding how sanitize is handled, so I have added reading those additional fields to output that other information.
The ZNR and no-deallocate bits are handled more-or-less the same way for any zoned device, so those bits are now handled in an anonymous union the same way.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>