make: Setting MSVC standard to C17 in meson.
Setting MSVC standard to C17 in meson.
This currently generates a warning since we are not using the built-in option to set the standard.
Currently we are still supporting some versions of MSVC that cannot specify the C11 or later standards (VS2015 & VS2017).
For now, accepting this warning and setting this so we can start working to transition to a newer C standard in the future.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>