bug: Fixing arbitration burst output for NVMe devices
Fixing the arbitration burst output as it was missing the most significant bit for this field to output.
I converted all outputs to using the M_GETBITRANGE macro to make this more consistent and easier to read and see that it is reading the correct range of bits for each field.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>