feat: Converting array declarations to DECLARE_ZERO_INIT_ARRAY macro
Switching all arrays declared on the stack to using this macro to better ensure they are zero initialized for all of their memory.
The old method type array[size] = { 0 } may only initialize the first member to zero, but this new macro will use a method to ensure it is completely zero initialized.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>