bug: Fixing erase max LBA restore logic
Fixing some of the logic in restoring a MaxLBA for any ATA method (HPA/AMAC/DCO) to take into account whether the device is already at maxLBA or not. This reduces some commands and can stop false failures from being reported when using this to restore the native maxLBA to the highest possible value.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>