bug: Fixing setting internal EPC struct for SAS devices causing invalid output
Fixing how the VPD and Mode pages are read for EPC drives. Standby Y and Z were mixed up in a bitfield which resulted in incorrect output.
There was a second bug where the mode page header was not taken into account so the standby timer values were not correctly accessed and causing invalid results.

[Seagate/openSeaChest#155]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>