feat: Converting M_GETBITRANGE to the appropriate function for the input and output types
Converting M_GETBITRANGE to the correct function to read a bit range and output it. This removes some extra turnary operators that are in the macro which may or may not get optimized out.
This also ensures consistent behavior for what is intended in the code in case the macro did not figure out the correct thing to do.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>