quick: Implementing a workaround for weird case when creating a psuedo unc
Found a drive that supports both the write uncorrectable command and the old read long/write long methods.
For some odd reason it always fails the write uncorrectable, but the read long/write long works as expected.
So this workaround switches to read long/write long in event of a failure and turns off the write uncorrectable command as it continues creating defects.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>