feat: Adding set timestamp, reset device statistics, and other cleanup
Adding function to set the date and time timestamp and updating how it is displayed to convert it to a more human readable timestamp.

Adding a function to reset/reinitialize supported device statistics.

Cleaned up more about how a statistic is looked up from its page and offset to be more readable and maintainable.

[Seagate/openSeaChest#172]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>