feat: Adding initial code for SATA CDL device statistics
Adding CDL device statistics on SATA drives.
This code supports concurrent ranges 0-3 per ACS-6 statistics.
This also handles the difference between whole device policies and concurrent range policies depending on what the drive populates when it is read.

All these statistics are defined in the spec as supporting the read then initialize feature as well according to the standards.

[Seagate/openSeaChest#172]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>