bug: Fixing incorrect logic on wue and read-write long retry
The logic for when to retry from write uncorrectable to the older read long/write long method was incorrect and causing drives that properly supported write uncorrectable to exit in a successful case.
The else to set "not supported" was in the wrong location and should have been before the retry to read long/write long.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>