bug: Fixing set timestamp and support power on milliseconds in date and time timestamp
Fixing a bug in the set date and time timestamp command for ATA drives.
It was checking the wrong byte when validating if the timestamp is supported.

When printing the date and time timestamp it is possible for it to be representing the power on time in milliseconds, so this also adds support for that.

[Seagate/openSeaChest#172]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>