quick: Fixing warning about uninitialized parameter in CDL code
Fixing a warning about uninitialized data. This was due to the parameter being marked as RO when it should have been RW.
The reason RW is needed is safe_strcat and safe_strlen functions will read the string data to determine length and where to concatenate so this should have been read-write.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>