bug: Fixing incorrect check for null when trim is run for NVMe devices in Windows.
There is a special case for Windows due to how NVMe is handled and when that function is run.
It reads a VPD page to confirm the unmap limits for Windows so the command(s) can be generated correctly.
There was a bug where after the allocation it was checking for != M_NULLPTR when it should have been == M_NULLPTR.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>