bug: Fixing FARM SOBR/SWR config output and physical element status.
Fixing the output for the SOBR/SWR config since it was outputting the same value for both due to accessing the wrong field in a structure.

Fixed the get physical element status output. It reports health and a timestamp so this also simplifies that field output as well.

A few other field names have been shortened to keep things nicely aligned in the output as well.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>