bug: Adding check after changing block descriptor to make sure the change is what was requested
There are some drives that will accept unsupported blocksizes or unsupported maxlba changes without generating an error message.
This adds an additional check to confirm that the block descriptor was updated as requested before continuing onwards.

In the case of maxlba changes, if the change did not happen, NOT_SUPPORTED is returned.
In the case of format unit, if the change did not happen to maxlba, a warning message is printed and if a change did not happen to the block size, it reverts to the original blocksize of the device.

[Seagate/openSeaChest#156]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>