bug: Fixing incorrect interpretation of the result of requesting supported opcode info for secprot
Fixing incorrect check of when SCSI security protocol commands are blocked causing erroneous output.
If we failed to discover supported security protocols, then later find the drive says it supports the command, then we know the command is blocked. Otherwise security protocol is not supported by the drive and the inability to read the protocol list is not actually an error.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>