make: Updating clang-tidy configuration
updating the clang-tidy config file for new features.
Some are things we should come back to enable, and others are disabling things we are not ready to support just yet.
We are disabling bugprone assignment in condition since we use that in a few places heavily and will require a big refactor.
We are disabling a modernization for macro to enum since that wants to use annonymous enums which is a C11 standard feature and we are not quite ready to adopt C11.
Finally disabling a include since we do handle some system oddness with our own files we include for portability and this may not otherwise solve the issue it is warning about.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>